Kubernetes_Assistant/requirements.txt
crewai==0.86.0
crewai_tools==0.17.0
langchain_community==0.3.5
langchain_ibm==0.3.2
#ibm-watsonx-ai==1.1.24
ibm-watsonx-ai
blinker==1.7.0
Bootstrap-Flask==2.3.3
cachetools==5.3.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
dominate==2.9.1
email-validator==2.1.0.post1
filelock==3.13.1
Flask==3.0.0
Flask-CKEditor==0.5.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-auth==2.27.0
greenlet==3.0.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.2
#kubernetes==29.0.0
kubernetes
MarkupSafe==2.1.3
oauthlib==3.2.2
ollama==0.4.4
openshift==0.13.2
pipenv==2023.11.15
platformdirs==4.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-string-utils==1.0.0
PyYAML==6.0.1
pydantic==2.10.3
#requests==2.32.0
requests
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
secret-key-generator==0.0.8
six==1.16.0
typing_extensions==4.12.2
#urllib3==2.2.2
urllib3
virtualenv==20.25.0
visitor==0.1.3
websocket-client==1.7.0
Werkzeug==3.0.1
WTForms==3.1.2
